requirements.txt
streamlit
autogen
pandas
numpy
matplotlib
openpyxl
python-dotenv
